Embedded auth into the labeling app. Left to test and verify working as expected. Also refactor
