--- labelapp/auth_app/README.md
@@ -1,4 +1,0 @@
-# Auth App
-
-Microservice used for authentication/authorization.
-Most likely will be using FastAPI.
