--- labelapp/auth_app/pyproject.toml
@@ -1,23 +1,0 @@
-[tool.poetry]
-name = "auth-app"
-version = "0.1.0"
-description = ""
-authors = ["sadelcarpio <sadelcarpioa@gmail.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-fastapi = "^0.109.2"
-uvicorn = {extras = ["standard"], version = "^0.27.1"}
-python-multipart = "^0.0.9"
-python-jose = {extras = ["cryptography"], version = "^3.3.0"}
-passlib = {extras = ["bcrypt"], version = "^1.7.4"}
-sqlmodel = "^0.0.14"
-pydantic = {extras = ["email"], version = "^2.6.1"}
-pydantic-settings = "^2.1.0"
-pg8000 = "^1.30.4"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
