+++ labelapp/labeling_app/app/api/deps.py
@@ -1,11 +1,9 @@
 from typing import Generator, Annotated
 
 from fastapi import Depends
-from fastapi import HTTPException
-from sqlmodel import Session, select
+from sqlmodel import Session
 
 from app.db.engine import EngineDep
-from app.models.users import UserModel
 
 
 def get_db(engine: EngineDep) -> Generator:
@@ -14,24 +12,3 @@ def get_db(engine: EngineDep) -> Generator:
 
 
 SessionDep = Annotated[Session, Depends(get_db)]
-
-
-# TODO: change with actual user authentication
-def get_current_user(session: SessionDep) -> UserModel:
-    user = session.exec(select(UserModel).where(UserModel.is_admin == False)).first()
-    if user is None:
-        raise HTTPException(status_code=404, detail="Current user not found.")
-    return user
-
-
-# TODO: change with actual admin user auth
-def get_current_admin_user(session: SessionDep) -> UserModel:
-    user = session.exec(select(UserModel).where(UserModel.is_admin == True)).first()
-    if user is None:
-        raise HTTPException(status_code=404, detail="Current admin user not found.")
-    return user
-
-
-# TODO: Have one dep to verify token and two deps to check user, and admin user
-CurrentUser = Annotated[UserModel, Depends(get_current_user)]
-CurrentAdminUser = Annotated[UserModel, Depends(get_current_admin_user)]
