+++ labelapp/labeling_app/app/schemas/token.py
@@ -1,0 +1,11 @@
+from sqlmodel import SQLModel
+
+
+class Token(SQLModel):
+    access_token: str
+    token_type: str
+
+
+class TokenData(SQLModel):
+    username: str | None = None
+    scopes: list[str] = []
