+++ labelapp/labeling_app/app/schemas/users.py
@@ -1,4 +1,5 @@
-from pydantic import EmailStr
+from pydantic import EmailStr, field_validator
+from pydantic_core.core_schema import ValidationInfo
 from sqlmodel import SQLModel, Field, AutoString
 
 
@@ -17,6 +18,18 @@ class UserRead(SQLModel):
 class UserCreate(UserBase):
     full_name: str
     password: str
+
+    @field_validator("password")
+    @classmethod
+    def validate_password(cls, password: str, values: ValidationInfo) -> str:
+        username = values.data.get("username"),
+        if password in username:
+            raise ValueError("Password can't contain username information.")
+        if len(password) < 8:
+            raise ValueError("Password must be at least 8 characters.")
+        if len(password) > 20:
+            raise ValueError("Password must be at most 20 characters.")
+        return password
 
 
 class UserUpdate(UserBase):
