+++ labelapp/labeling_app/app/security.py
@@ -1,0 +1,95 @@
+from datetime import timedelta, datetime, timezone
+from typing import Annotated
+
+from fastapi import Depends, Security, HTTPException
+from fastapi import status
+from fastapi.security import OAuth2PasswordBearer, SecurityScopes
+from jose import jwt, JWTError
+from passlib.context import CryptContext
+from pydantic import ValidationError
+from sqlmodel import select
+
+from app.api.deps import SessionDep
+from app.models.users import UserModel
+from app.schemas.token import TokenData
+
+SECRET_KEY = "546fc0304773380edc8f1171636ceaf47f9660ad54f26d7a3c935daf645ee5a3"
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token",
+                                     scopes={"admin": "Admin functionalities"})
+
+
+def verify_password(plain_password, hashed_password):
+    return pwd_context.verify(plain_password, hashed_password)
+
+
+def get_password_hash(password):
+    return pwd_context.hash(password)
+
+
+# Temporarily then pass it to CRUDUser
+def get_user(session: SessionDep, username: str):
+    user = session.exec(select(UserModel).where(UserModel.username == username)).first()
+    return user
+
+
+def authenticate_user(session: SessionDep, username: str, password: str):
+    user = get_user(session, username)
+    if not user:
+        return False
+    if not verify_password(password, user.hashed_password):
+        return False
+    return user
+
+
+def create_access_token(data: dict, expires_delta: timedelta | None = None):
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.now(timezone.utc) + expires_delta
+    else:
+        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+    return encoded_jwt
+
+
+TokenDep = Annotated[TokenData, Depends(oauth2_scheme)]
+
+
+async def get_current_user(security_scopes: SecurityScopes, token: TokenDep, session: SessionDep):
+    if security_scopes.scopes:
+        authenticate_value = f'Bearer scope="{security_scopes.scope_str}"'
+    else:
+        authenticate_value = "Bearer"
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Could not validate credentials",
+        headers={"WWW-Authenticate": authenticate_value},
+    )
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+        token_scopes = payload.get("scopes", [])
+        token_data = TokenData(scopes=token_scopes, username=username)
+    except (JWTError, ValidationError):
+        raise credentials_exception
+    user = get_user(session, token_data.username)
+    if user is None:
+        raise credentials_exception
+    for scope in security_scopes.scopes:
+        if scope not in token_data.scopes:
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail="Not enough permissions",
+                headers={"WWW-Authenticate": authenticate_value},
+            )
+    return user
+
+
+CurrentUser = Annotated[UserModel, Depends(get_current_user)]
+CurrentAdminUser = Annotated[UserModel, Security(get_current_user, scopes=["admin"])]
