+++ labelapp/labeling_app/pyproject.toml
@@ -14,6 +14,9 @@ pg8000 = "^1.30.4"
 pydantic-settings = "^2.1.0"
 pydantic = {extras = ["email"], version = "^2.5.3"}
 httpx = "^0.26.0"
+passlib = {extras = ["bcrypt"], version = "^1.7.4"}
+python-jose = {extras = ["cryptography"], version = "^3.3.0"}
+python-multipart = "^0.0.9"
 
 
 [build-system]
