+++ labelapp/labeling_app/tests/test_projects.py
@@ -7,7 +7,8 @@ from fastapi.testclient import TestClient
 from sqlmodel import Session
 
 from app import create_app
-from app.api.deps import get_db, get_current_user
+from app.api.deps import get_db
+from app.security import get_current_user
 from app.core.config import Settings, get_settings
 from app.crud import CRUDUser
 from app.crud.crud_project import get_projects_crud, CRUDProject
