+++ labelapp/labeling_app/tests/test_urls.py
@@ -6,7 +6,8 @@ from fastapi.testclient import TestClient
 from sqlmodel import Session
 
 from app import create_app
-from app.api.deps import get_db, get_current_user
+from app.api.deps import get_db
+from app.security import get_current_user
 from app.core.config import Settings, get_settings
 from app.models.extras import UserProjectModel
 from app.models.urls import UrlModel
